check that SpeX wavelength grids are actually at the same R
Because I need to do R-dependent smoothing, I've created a model grid
that is smoothed to the R-array of ONE SpeX spectrum, but if I'm going
to apply that to all the other BD spectra from SpeX I need to know that
the R vs. Lambda curves are the same for all the SpeX spectra.  So this
plots R vs. lambda for all 12 of my SpeX spectra.