change interpolation check in make_model
the interpolation check was focused on the length of the array, but
that's not enough of a check. Now it checks that the first
elements of the wavelength arrays are the same, and if not it
will interpolate (this should be fine, since the R vs. lambda curve
is consistent across objects)