debugged plots, commented out quantile plot, added errorbars to random
Was having a problem adding errorbars to the plot, fixed inputs.
Commented quantile plot because it was throwing an error and I
didn't feel like dealing with it.
Changed the format of self.all_params from np array to list, because
emcee_plot couldn't handle it as an array.
Added errorbars to the plot of random draws, representing the data
uncertainty added in quadrature to the additional error from the fit.