plot best-fit ln(s), because errorbars don't obey the alpha command
plot the ln(s) value corresponding to the 50th quantile, but plot
it on the actual data.  It doesn't work to plot it on the random
draw models because the error bars don't use the alpha command, so they are all opaque and you can't see the model variations very well.