plot random set of models, including uncertainty
just like the function that's still in bdfit.BDSampler, picks a
random set of models to plot, and plots them, including uncertainty.
Splitting to a new module allows plotting without rerunning the fit,
and allows other plotting codes to access it.

Eventually I might want to make this universal enough that it can
work for any kind of model. Maybe.